Updates based on review
