+++ tasks/src/robocorp/tasks/cli.py
@@ -9,8 +9,36 @@ Note: when running tasks, clients using this approach MUST make sure that any
 code which must be automatically logged is not imported prior the the `cli.main`
 call.
 """
+import os
 import sys
+import warnings
 from typing import List, Optional, Protocol
+
+
+def inject_truststore():
+    # Use certificates from native storage (if `truststore` installed)
+
+    # RC_USE_TRUSTSTORE env var is only used to avoid unwanted warnings in tests
+    # and should not be used for other purposes
+    if os.getenv("RC_USE_TRUSTSTORE").lower() in ["false", "0"]:
+        return
+
+    if sys.version_info >= (3, 10):
+        try:
+            import truststore  # type: ignore
+
+            truststore.inject_into_ssl()
+        except ModuleNotFoundError:
+            warnings.warn(
+                "Usage of the native system certificate stores can’t be enabled, ensure you have the"
+                " `robocorp-truststore` dependency installed in the environment.",
+                Warning,
+                stacklevel=2,
+            )
+            pass
+
+
+inject_truststore()
 
 if sys.platform == "win32":
     # Apply workaround where `asyncio` would halt forever when windows UIAutomation.dll
@@ -20,9 +48,7 @@ if sys.platform == "win32":
     sys.coinit_flags = _COINIT_MULTITHREADED  # type:ignore
 
 # Just importing is enough to register the commands
-from . import _commands, inject_truststore  # noqa
-
-inject_truststore()
+from . import _commands  # noqa
 
 
 class IArgumentsHandler(Protocol):
